add and setup vue-router
