+++ package.json
@@ -20,6 +20,7 @@
     },
     "dependencies": {
         "laravel-vue-i18n": "^1.2.0",
-        "vue": "^3.2.29"
+        "vue": "^3.2.29",
+        "vue-router": "^4.0.12"
     }
 }
