+++ resources/js/views/Detail.vue
@@ -1,0 +1,17 @@
+<template>
+    <div class="border-black border-t-2">
+        <h1 class="p-5 text-2xl">Detail diela</h1>
+    </div>
+    <div class="bg-black text-white h-full">
+        <div class="p-5">
+            <div class="text-xl mb-4">{{ $t('Collect artworks on your visit') }}</div>
+            <div>{{ $t('View them later from home. Enter code') }}</div>
+        </div>
+    </div>
+</template>
+
+<script>
+    export default {
+
+    }
+</script>
